% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Sum of two vectors of numbers}
\usage{
add(x, y, negative = FALSE)
}
\arguments{
\item{x}{a vector of numbers}

\item{y}{a vector of numbers}

\item{negative}{logical, whether to flip the sign of the sum}
}
\value{
the sum of the two vectors
}
\description{
This function sums two vectors of numbers and optionally
allows for the negative of the sum to be returned.
}
\examples{

add(1:5, 6:10)
add(1:5, 6:10, negative=TRUE)

}
